event tests updated except last one
